+++ src/replica/construct.rs
@@ -259,6 +259,7 @@ impl<K: Key + Hash, V: Value> Replica<K, V> {
                 bulk_in_flight: Arc::new(RwLock::new(HashSet::new())),
                 bulk_dumps_in_flight: Arc::new(AtomicUsize::new(0)),
                 max_concurrent_bulk_dumps: config.max_concurrent_bulk_dumps,
+                max_value_size: config.max_value_size,
                 broadcasts_in_flight: Arc::new(AtomicUsize::new(0)),
                 max_concurrent_broadcasts: config.max_concurrent_broadcasts,
                 pending_dumps: Arc::new(RwLock::new(HashMap::new())),